move all componets into components folder
