+++ src/App.js
@@ -1,4 +1,4 @@
-import Topbar from "./topbar/Topbar";
+import Topbar from '../src/components/topbar/Topbar'
 import Home from './pages/home/Home'
 
 function App() {
