+++ src/pages/home/Home.jsx
@@ -1,6 +1,7 @@
-import Header from '../../header/Header'
-import Posts from '../../posts/Posts'
-import Sidebar from '../../sidebar/Sidebar'
+import Header from '../../components/header/Header'
+import Posts from '../../components/posts/Posts'
+import Sidebar from '../../components/sidebar/Sidebar'
+
 import './Home.css'
 
 export default function Home() {
